Improve Build & Packaging Setup
Improved the build and packaging setup by migrating to the sole use of
`pyproject.toml` rather than a combination of `pyproject.toml`,
`setup.cfg`, `setup.py` and `MANIFEST.in`, simplifying and
consolidating the build and setup process.

Rearranged the `requirements.*.txt` files, separating the dependencies
into those needed by the library (`requests`), those needed for
development (`purest` and `black`) and those needed for building and
distribution (`build`, `twine` and `wheel`).

Incremented the library version to `1.1.6`.

+++ build.sh
@@ -1,12 +1,29 @@
 #!/bin/sh
+
 # See https://share.getty.edu/display/SOFTARCH/Using+the+Nexus+Package+Manager+with+Getty+Packages for how to configure twine to publish
-# Install the build requirements too! (twine, wheel)
-pip install -r build_requirements.txt
+
+# Run the unit tests, refusing to deploy if the unit tests fail
+if ! docker compose run tests; then
+	echo -e "The unit tests failed; cannot publish!";
+	exit 1;
+fi
+
+# Install the build requirements (twine, wheel)
+pip install -r requirements.build.txt
 
 rm -Rrf dist/*
 
-## build the final version
-python setup.py sdist bdist_wheel
+## Build the library
+if ! python -m build; then
+    echo -e "The build failed; cannot publish!";
+    exit 1;
+fi
+
+## Test the build is valid
+if ! twine check dist/*; then
+    echo -e "The build check tests failed; cannot publish!";
+    exit 1;
+fi
 
 ## upload to getty nexus registry
 twine upload --repository-url "https://artifacts.getty.edu/repository/jpgt-pypi/" dist/*
